[FIX] : 분석 거래 퍼센트 변수타입 double로 변경

+++ Backend/src/main/java/com/example/backend/domain/analyze/dto/AnalyzeResponseDto.java
@@ -14,7 +14,7 @@ public class AnalyzeResponseDto {
     public static class DataBody {
 
         private Long day;
-        private Long totalBudget;
+        private Double totalBudget;
         private Integer amountUsed;
 
         @JsonProperty("음식점")
